Add python3.5 support
Specifying the `levels` option brings python3.5 support. The meaning of this argument has valuably changed from python2 -> python3 as far as that by default discovery of the modules in the current directory is not working properly.
Specifying level = 1 tells the import module that it should search in the current directory, and thus that makes the code work on python3 again.